+++ GeekStore/GeekStore.UnitTest/FactoryFixture.cs
@@ -10,7 +10,7 @@ namespace GeekStore.UnitTests
         [TestCase(1.0, 2.0, CPUCores.DualCore, CPUManufacturer.Intel, "Core 2 Duo", "LGA775", 65, 2)]
         [TestCase(2.0, 1.0, CPUCores.DualCore, CPUManufacturer.Intel, "Core 2 Duo", "LGA775", 65, 2)]
 
-        public void TestCPUFactory(double baseFrequency, double boostFrequency, CPUCores cores, CPUManufacturer manufacturer, string model, string socket, int tdp, int threads)
+        public void TestCPUFactory(decimal baseFrequency, decimal boostFrequency, CPUCores cores, CPUManufacturer manufacturer, string model, string socket, int tdp, int threads)
         {
             CPU cpu = CPUFactory.CreateCPU(baseFrequency, boostFrequency, cores, manufacturer, model, socket, tdp, threads);
             Assert.IsInstanceOf(typeof(CPU), cpu);
